requirements.txt
litellm==1.81.12
python-dotenv==1.2.1
